Add files via upload
Just completed the second challenge on day 25! This one took a while to get working properly, so I spread the work over four days. This app is a game of rock, paper, scissors. The computer will generate a random move and tell the user if they should win or lose, then the user selects between the three choices. If the user is correct they will get a point, if they are wrong they will lose a point. The app will keep track of the users score, reset the game after 10 turns, and will display messages between rounds. I'm not fully done with this just yet, while the app is functional I will be going back in and updating it to have a better looking UI.